feat(01): Adding backend outbox service

+++ packages/authorization-server/src/main/java/com/clinton/authorization_server/service/AccountService.java
@@ -1,11 +1,11 @@
 package com.clinton.authorization_server.service;
 
+import com.clinton.authorization_server.exceptions.AccountNotFoundException;
 import com.clinton.authorization_server.model.Account;
 import com.clinton.authorization_server.repository.AccountRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.dao.DataAccessException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.TransactionException;
 import org.springframework.transaction.annotation.Isolation;
@@ -14,19 +14,14 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
 import java.sql.SQLException;
-import java.text.MessageFormat;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Optional;
 
 /**
  * @author Clinton Fernandes
  */
 @Service
+@Profile("auth")
 public class AccountService {
-
-    Map<String, String> success = new HashMap<>(1);
-    Map<String, String> failure = new HashMap<>(1);
     private final AccountRepository accountRepository;
 
     @Autowired
@@ -35,65 +30,40 @@ public class AccountService {
     }
 
     @Transactional(propagation = Propagation.REQUIRES_NEW, isolation = Isolation.SERIALIZABLE)
-    public ResponseEntity<Map<String, String>> createAccount(Account account) {
-        account = accountRepository.save(account);
-        success.put("Success", "Account created with account number: " + account.getAccountNumber());
-        return new ResponseEntity<>(success, HttpStatus.CREATED);
+    public Account createAccount(Account account) {
+        account.setAccountBalance(BigDecimal.ZERO);
+        account.setReservedAmount(BigDecimal.ZERO);
+        return accountRepository.save(account);
     }
 
     @Transactional(propagation = Propagation.REQUIRES_NEW, isolation = Isolation.SERIALIZABLE)
-    public ResponseEntity<Map<String, String>> deleteAccount(Integer accountId) {
-        accountRepository.delete(accountRepository.getReferenceById(accountId));
-        success.put("Success", "Account deleted with account number: " + accountId);
-        return new ResponseEntity<>(success, HttpStatus.CREATED);
+    public void deleteAccount(Integer accountId) {
+        if (accountRepository.existsById(accountId)) {
+            accountRepository.delete(accountRepository.getReferenceById(accountId));
+        } else {
+            throw new AccountNotFoundException("Account with id " + accountId + " not found");
+        }
     }
 
     @Transactional(propagation = Propagation.REQUIRES_NEW, isolation = Isolation.READ_COMMITTED)
-    public ResponseEntity<Map<String, String>> creditAccount(Integer accountId, BigDecimal amount) {
-        Optional<Account> optionalAccount = accountRepository.findById(accountId);
-        if (optionalAccount.isPresent()) {
-            optionalAccount.get().setAccountBalance(optionalAccount.get().getAccountBalance().add(amount));
-            accountRepository.save(optionalAccount.get());
-            success.put("Success", MessageFormat.format("Amount {0} credited to account {1}", amount, accountId));
-            return new ResponseEntity<>(success, HttpStatus.ACCEPTED);
+    public void creditAccount(Integer accountId, BigDecimal amount) {
+        Optional<Account> account = accountRepository.findById(accountId);
+        if (account.isPresent()) {
+            account.get().creditAccount(amount);
+            accountRepository.save(account.get());
         } else {
-            failure.put("Failure", "Account not found!");
-            return new ResponseEntity<>(failure, HttpStatus.NOT_FOUND);
+            throw new AccountNotFoundException("Account with id " + accountId + " not found");
         }
     }
 
     @Transactional(propagation = Propagation.REQUIRES_NEW, isolation = Isolation.READ_COMMITTED, rollbackFor = {DataAccessException.class, SQLException.class, TransactionException.class})
-    public ResponseEntity<Map<String, String>> debitAccount(Integer accountId, BigDecimal amount) {
-        Optional<Account> optionalAccount = accountRepository.findById(accountId);
-        Map<String, String> result = new HashMap<>();
-        if (optionalAccount.isPresent()) {
-            if (optionalAccount.get().getAccountBalance().subtract(amount).compareTo(BigDecimal.ZERO) < 0) {
-                failure.put("Failure", "Insufficient Funds!");
-                return new ResponseEntity<>(failure, HttpStatus.OK);
-            } else {
-                optionalAccount.get().setAccountBalance(optionalAccount.get().getAccountBalance().subtract(amount));
-                Account account = accountRepository.save(optionalAccount.get());
-                success.put("Success", MessageFormat.format("Account {0} debited with {1}, remaining balance: {2}", accountId, amount, account.getAccountBalance()));
-                return new ResponseEntity<>(success, HttpStatus.ACCEPTED);
-            }
+    public void debitAccount(Integer accountId, BigDecimal amount) {
+        Optional<Account> account = accountRepository.findById(accountId);
+        if (account.isPresent()) {
+            account.get().debitAccount(amount);
+            accountRepository.save(account.get());
         } else {
-            failure.put("Failure", "Account not found!");
-            return new ResponseEntity<>(failure, HttpStatus.NOT_FOUND);
-        }
-    }
-
-    @Transactional(readOnly = true, propagation = Propagation.REQUIRES_NEW, isolation = Isolation.READ_COMMITTED)
-    public boolean checkAccount(Integer accountId) {
-        return accountRepository.findById(accountId).isPresent();
-    }
-
-    @Transactional(readOnly = true, propagation = Propagation.REQUIRED, isolation = Isolation.READ_COMMITTED)
-    public Account findAccount(Integer accountId) {
-        Optional<Account> optionalAccount = accountRepository.findById(accountId);
-        if (optionalAccount.isPresent()) {
-            return optionalAccount.get();
-        } else {
-            throw new RuntimeException("Account Id: " + accountId + " not found");
+            throw new AccountNotFoundException("Account with id " + accountId + " not found");
         }
     }
 }
